Fix SelectionCoordinator’s strong reference to the presentingViewController.
Additionally cleaned up extensions in Core.
