+++ Core/Source/Extensions/ReactiveSwift/UIKit/UIViewController.swift
@@ -72,4 +72,28 @@ extension Reactive where Base: UIViewController {
                     .filter { $0.view.window == nil }
             }
     }
+
+    public var present: Action<(UIViewController, Bool), (), NoError> {
+        let baseViewController = base
+        return Action { (viewController, animated) in
+            return SignalProducer { [weak baseViewController] (observer, _) in
+                baseViewController?.present(viewController, animated: animated, completion: {
+                    observer.sendCompleted()
+                })
+            }
+
+        }
+    }
+
+    public var dismiss: Action<Bool, (), NoError> {
+        let baseViewController = base
+        return Action { animated in
+            return SignalProducer { [weak baseViewController] (observer, _) in
+                baseViewController?.dismiss(animated: animated, completion: {
+                    observer.sendCompleted()
+                })
+            }
+
+        }
+    }
 }
